colormaps, walllights, etc now bytes instead of offsets. offsets calculated/lookued up later (they are * 256). some shift 4 lookups removed. made a generic drawcolumn prep pair of calls, may combine both into one soon

+++ r_defs.h
@@ -472,7 +472,7 @@ typedef struct vissprite_s
 
     // for color translation and shadow draw,
     //  maxbright frames as well
-    uint16_t	colormap;
+    uint8_t	colormap;
    
     int32_t			mobjflags;
